+++ core/src/main/kotlin/net/corda/core/flows/StartableByRPC.kt
@@ -1,0 +1,14 @@
+package net.corda.core.flows
+
+import java.lang.annotation.Inherited
+import kotlin.annotation.AnnotationTarget.CLASS
+
+/**
+ * Any [FlowLogic] which is to be started by the RPC interface ([net.corda.core.messaging.CordaRPCOps.startFlowDynamic]
+ * and [net.corda.core.messaging.CordaRPCOps.startTrackedFlowDynamic]) must have this annotation. If it's missing the
+ * flow will not be allowed to start and an exception will be thrown.
+ */
+@Target(CLASS)
+@Inherited
+@MustBeDocumented
+annotation class StartableByRPC